rota para selecionar ganhador da partida

+++ database/migrations/2023_01_23_061106_create_championships_players_table.php
@@ -19,7 +19,7 @@ class CreateChampionshipsPlayersTable extends Migration
             $table->integer('id_player');
             $table->float('points')->default(0);
             $table->integer('defeats')->default(0);
-            $table->integer('ppm')->default(1);
+            $table->float('ppm')->default(1);
             $table->timestamps();
         });
     }
